refactor(StatusBuilder) : 불변 객체 유지를 위한 Builder 구현
